v0.15 - Changes to id generation and parsing;

+++ models/order.go
@@ -3,10 +3,10 @@ package models
 import "time"
 
 type Order struct {
-	ID           uint      `json:"id" gorm:"primaryKey"`
-	ProductRefer uint      `json:"product_id"`
+	ID           uint64    `json:"id" gorm:"primaryKey"`
+	ProductRefer uint64    `json:"product_id"`
 	Product      Product   `gorm:"foreignKey:ProductReferer"`
-	UserReferer  uint      `json:"user_id"`
+	UserReferer  uint64    `json:"user_id"`
 	User         User      `gorm:"foreignKey:UserReference"`
 	CreatedAt    time.Time `gorm:"autoCreateTime"`
 }
